+++ Engine/Common/AiService.cs
@@ -30,6 +30,21 @@ namespace JocysCom.VS.AiCompanion.Engine
 		public bool IsDefault { get => _IsDefault; set => SetProperty(ref _IsDefault, value); }
 		bool _IsDefault;
 
+		/// <summary>Stream response data as it becomes available.</summary>
+		[DefaultValue(true)]
+		public bool ResponseStreaming { get => _ResponseStreaming; set => SetProperty(ref _ResponseStreaming, value); }
+		bool _ResponseStreaming;
+
+		/// <summary>Stream response data as it becomes available.</summary>
+		[DefaultValue(600)]
+		public int ResponseTimeout { get => _ResponseTimeout; set => SetProperty(ref _ResponseTimeout, value); }
+		int _ResponseTimeout;
+
+		/// <summary>Configure for Microsoft Azure OpenAI.</summary>
+		[DefaultValue(false)]
+		public bool IsAzureOpenAI { get => _IsAzureOpenAI; set => SetProperty(ref _IsAzureOpenAI, value); }
+		bool _IsAzureOpenAI;
+
 		#region API Keys
 
 		/// <summary>Organization key. Usage from these API requests will count against the specified organization's subscription quota.</summary>