+++ Engine/Companions/ChatGPT/base_completion_response.cs
@@ -1,0 +1,12 @@
+﻿namespace JocysCom.VS.AiCompanion.Engine.Companions.ChatGPT
+{
+	public class base_completion_response: base_item
+	{
+		public string id { get; set; }
+		public string @object { get; set; }
+		public int created { get; set; }
+		public string model { get; set; }
+		public usage usage { get; set; }
+
+	}
+}