+++ Engine/Companions/ChatGPT/base_item.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections.Generic;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace JocysCom.VS.AiCompanion.Engine.Companions.ChatGPT
+{
+	public class base_item
+
+	{
+		/// <summary>
+		/// This property is used to capture additional JSON properties that aren't explicitly defined in the class.
+		/// These are stored as KeyValue pairs, where the 'Key' is the property name and 'Value' is the property value.
+		/// </summary>
+		[JsonExtensionData]
+		public Dictionary<string, JsonElement> additional_properties
+		{
+			get => _additional_properties ?? (_additional_properties = new Dictionary<string, JsonElement>());
+			set => _additional_properties = value;
+		}
+		private Dictionary<string, JsonElement> _additional_properties;
+
+	}
+}