+++ Engine/Companions/ChatGPT/chat_completion_response.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace JocysCom.VS.AiCompanion.Engine.Companions.ChatGPT
+{
+	internal class chat_completion_response: base_completion_response
+	{
+		public List<chat_choice> choices { get; set; }
+	}
+}