+++ Engine/Companions/ChatGPT/message_role.cs
@@ -1,0 +1,10 @@
+﻿namespace JocysCom.VS.AiCompanion.Engine.Companions.ChatGPT
+{
+	public enum message_role
+	{
+		system = 0,
+		user = 1,
+		assistant = 2,
+		function = 3,
+	}
+}