+++ Engine/Companions/ChatGPT/model.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace JocysCom.VS.AiCompanion.Engine.Companions.ChatGPT
+{
+	public partial class model: base_item
+	{
+		public string id { get; set; }
+
+		public string @object { get; set; }
+
+		public DateTime created { get; set; }
+
+		public string owned_by { get; set; }
+
+	}
+
+}