+++ Engine/Companions/ChatGPT/usage.cs
@@ -1,0 +1,12 @@
+﻿namespace JocysCom.VS.AiCompanion.Engine.Companions.ChatGPT
+{
+	public class usage : base_item
+	{
+		public int prompt_tokens { get; set; }
+
+		public int completion_tokens { get; set; }
+
+		public int total_tokens { get; set; }
+
+	}
+}