+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+2023-09-07 v1.1.6
+
+- Fixed: Task cloning, stop sending.
+- New: Added support for streaming replies.
+- New: Added service timeout option.
+- New: Warning when API URL is insecure and not local.
+- Update: Improved support for Azure OpenAI.
+- Update: Improved support for GPT4All.
+
 2023-08-31 v1.0.115
 
 - Fix: Unable to hide template or task list panel.